Merge client PR check and deploy workflows
These workflows are pretty redundant (they do the same thing except one deploys at the end and run in response to different events). That duplication makes it easy for things to get out of sync and make mistakes. This merges the two workflows into one, and makes the deploy step conditional, so it only runs on pushes to `main`, whereas the rest also run on pull requests.